# Interpolate '' instead of nil when multiple is false. Part two!