@matthewd 2018-03-21
As you noted, this code is not part of the regular request path, so I don't think it's worth bothering with here (or anywhere else similar, that doesn't expect multiple calls per request).

It does seem worth looking at upstream, though: have you reported this behaviour to MRI?