@oniofchaos 2018-03-21
I feel like just because it's not part of every rails request doesn't mean it can't be improved. If that was the case then why is bootsnap being added by default in 5.2?

If the code after the change is confusing I'm definitely open to tweaking it. 

Can you elaborate on what behavior you're suggesting I report to MRI?

Thanks for the feedback.