@matthewd 2018-03-21
> why is bootsnap being added

Bootsnap makes a measurable difference to app boot time; a literal microbenchmark, which shows allocating two objects is 1.015µs slower than allocating one, is not that.

We allocate quite a few objects, especially during boot. Eliminating hotspots is always worthwhile, but chasing individual allocations is not.

> For my app, this saved 3698 allocations of "" - around 150 KB of savings

Is 3.75ms a notable portion of your application's boot time, or first request time? A real-world demonstration that this actually makes a difference might be more compelling.

This is not "150 KB of savings" -- that suggests it reduces the total memory requirements of a running application; it's 150 KB of objects the GC doesn't need to allocate and then almost immediately discard. That takes _non-zero_ effort, but it's something the GC is pretty good at.

> Can you elaborate on what behavior you're suggesting I report to MRI?

The fact that `x = nil; ".#{x}."` allocates two objects when `NilClass#to_s` has not been redefined.