@oniofchaos 2018-03-23
Bootsnap creates a larger difference, and it's also a more substantial addition of code. IMO shaving off thousands of allocations is a good improvement for a one line change. An ocean is made of a multitude of drops.

Also, to me any change that lessens the work that GC has to do while not otherwise compromising speed is a good change. 

You're right about the phrasing re: "150 KB of savings" being misleading based on GC behavior. Based on your feedback about practical demonstration, I booted up my app with Rails master vs Rails with this change. It was hard to pinpoint an exact amount of difference as memory usage fluctuated by small amounts but with the change it was generally lower by a few hundred KB (226.0 MB on master and 225.6 to 225.7 MB with the change). 

I filed a bug ticket here: https://bugs.ruby-lang.org/issues/14624. Even still, I think it's something worth fixing as not every user will always be able to get on the latest Ruby.